requirements.txt
fastapi
uvicorn[standard]
sqlmodel
asyncpg
opencv-python
python-dotenv